requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
serpapi
fpdf2

